requirements.txt
langgraph>=0.2.0 
langchain>=0.2.0 
langchain_openai>=0.1.0 
python-dotenv>=1.0.0 
graphviz>=0.20.1 
scikit-learn>=1.2.2
rich>=13.7.1